pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "psn-api"
version = "0.1.0"
description = "PlayStation Network API"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "psnawp-api",
    "python-dotenv",
]

[tool.setuptools]
package-dir = {"" = "."}
packages = ["src"]